requirements.txt
moderngl
pillow
pygame
numpy
